requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.8
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.4
idna==2.8
kombu==4.6.3
Pillow==6.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
qiniu==7.2.4
redis==3.2.1
requests==2.22.0
urllib3==1.25.3
vine==1.3.0
